Move #2

+++ Pixel-Adventure/Assets/Scripts/Packet/ServerSession.cs
@@ -15,7 +15,7 @@ public class ServerSession : PacketSession
 		MsgId msgId = (MsgId)Enum.Parse(typeof(MsgId), msgName);
 		ushort size = (ushort)packet.CalculateSize();
 		byte[] sendBuffer = new byte[size + 4];
-		Array.Copy(BitConverter.GetBytes(size + 4), 0, sendBuffer, 0, sizeof(ushort));
+		Array.Copy(BitConverter.GetBytes((ushort)(size + 4)), 0, sendBuffer, 0, sizeof(ushort));
 		Array.Copy(BitConverter.GetBytes((ushort)msgId), 0, sendBuffer, 2, sizeof(ushort));
 		Array.Copy(packet.ToByteArray(), 0, sendBuffer, 4, size);
 		Send(new ArraySegment<byte>(sendBuffer));
